refactor(adaptive-admin): calibration / mock / certificate as tabs, not cards
Replaces the launcher-card grid on the adaptive course detail page with pills
beside Content · Students · Cover art:

- Calibration tab consolidates everything calibration — schedule + assessment
  (CalibrationAdminSection) AND the per-student results (CalibrationResultsSection).
- Mock interviews tab — the calibration interview, rounds, and attempts.
- Certificate tab — template upload + unlock criteria.

Each sub-page body is extracted into a reusable section component
(CertificateAdminSection, MockInterviewAdminSection, CalibrationResultsSection);
the standalone /calibration, /mock-interview, /certificate routes become thin
wrappers around the same sections, so deep links keep working with no duplicated
logic.
